refactor: support float-based damage system - Enemy health changed to float for precise damage values - EnemyPrototype interface updated with float getHealth() - Boss also supports float damage Enables balanced weapon strategies with fractional damage

+++ core/src/main/java/game/mygame/entities/Boss.java
@@ -7,8 +7,8 @@ import com.badlogic.gdx.math.Rectangle;
 public class Boss {
     protected float x, y;
     protected float speed;
-    protected int health;
-    protected int maxHealth;
+    protected float health;
+    protected float maxHealth;
     protected int scoreValue;
     protected final float width, height;
     protected final Texture texture;
@@ -52,7 +52,7 @@ public class Boss {
         }
     }
 
-    public void hit(int damage) {
+    public void hit(float damage) {
         health -= damage;
         if (health <= 0) {
             alive = false;
